Maximum Drawdown (MDD)
In trading, maximum drawdown (MDD) is the maximum loss observed from a portfolio's peak to trough before a new peak is reached. It's calculated as the difference between the value of the lowest trough and the highest peak before the trough. MDD is expressed as a percentage value and is calculated over a long time period when the value of an asset or investment has gone through several boom-bust cycles.